+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.5.0
-*
-*
+
+* Add support for Python `3.10` version
+* Add command aliases